Move dialog to Application project
