+++ sources/Google.Solutions.IapDesktop.Application/Google.Solutions.IapDesktop.Application.csproj
@@ -161,6 +161,13 @@
     <Compile Include="Util\TypographicQuotes.cs" />
     <Compile Include="Util\Filename.cs" />
     <Compile Include="Views\About\AboutViewModel.cs" />
+    <Compile Include="Views\Authorization\AuthorizeView.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Views\Authorization\AuthorizeView.Designer.cs">
+      <DependentUpon>AuthorizeView.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Views\Authorization\AuthorizeViewModel.cs" />
     <Compile Include="Views\Authorization\DeviceFlyoutViewModel.cs" />
     <Compile Include="Views\Authorization\NewProfileView.cs">
       <SubType>Form</SubType>
@@ -385,6 +392,9 @@
     <EmbeddedResource Include="Views\About\AboutView.resx">
       <DependentUpon>AboutView.cs</DependentUpon>
     </EmbeddedResource>
+    <EmbeddedResource Include="Views\Authorization\AuthorizeView.resx">
+      <DependentUpon>AuthorizeView.cs</DependentUpon>
+    </EmbeddedResource>
     <EmbeddedResource Include="Views\Authorization\DeviceFlyoutView.resx">
       <DependentUpon>DeviceFlyoutView.cs</DependentUpon>
     </EmbeddedResource>
@@ -487,6 +497,12 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Commands\" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Resources\LogoWatermark_180.png" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Resources\GoogleSignIn_24.png" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\MSBuild.AssemblyVersion.1.3.0\build\MSBuild.AssemblyVersion.targets" Condition="Exists('..\packages\MSBuild.AssemblyVersion.1.3.0\build\MSBuild.AssemblyVersion.targets')" />
