+++ sources/Google.Solutions.IapDesktop.Application/Properties/Resources.resx
@@ -127,9 +127,6 @@
   <data name="logo" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\logo.ico;System.Drawing.Icon, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
-  <data name="AccentGradient_450" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\AccentGradient_450.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
   <data name="Computer_16" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\Computer_16.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
@@ -144,5 +141,107 @@
   </data>
   <data name="Refresh_16" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\Refresh_16.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="AccentGradient_450" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\AccentGradient_450.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="AuthorizationSuccessful" xml:space="preserve">
+    <value>&lt;!DOCTYPE html&gt;
+&lt;html&gt;
+    &lt;head&gt;
+        &lt;title&gt;Authorization successful&lt;/title&gt;
+        &lt;style&gt;
+            body {
+                font-family: 'Roboto', 'Helvetica', 'Arial', sans-serif;
+                font-size: 14px;
+                color: #5f6368;
+                text-align: center;
+                
+                
+                background-color: #F5F5F5;
+                color: black;
+            }
+            
+            h1 {
+                font-size: 24px;
+                font-weight: normal;
+                color: #6288F2;
+                font-weight: 400;
+            }
+            
+            .container-outer {
+                display: table;
+                position: absolute;
+                top: 0;
+                left: 0;
+                height: 100%;
+                width: 100%;
+            }
+
+            .container-middle {
+                display: table-cell;
+                vertical-align: middle;
+            }
+
+            .container-content {
+                margin-left: auto;
+                margin-right: auto;
+                width: 420px;
+                padding: 30px;
+                border: 1px solid lightgray;
+                background-color: white;
+            }
+            
+            /* Dark mode */
+            @media (prefers-color-scheme: dark) {
+                body {
+                    background-color: #555555;
+                    color: #F1F1F1;
+                }
+                
+                .container-content {
+                    background-color: #2D3031;
+                    border: 0px;
+                }
+            }
+        &lt;/style&gt;
+    &lt;/head&gt;
+
+    &lt;body&gt;
+        &lt;div class="container-outer"&gt;
+            &lt;div class="container-middle"&gt;
+                &lt;div class="container-content" id="success"&gt;
+                    &lt;h1&gt;&lt;span class='product'&gt;IAP Desktop&lt;/span&gt; is ready to use&lt;/h1&gt;
+                    &lt;p&gt;
+                        You can now use IAP Desktop to access your VMs.
+                    &lt;/p&gt;
+                &lt;/div&gt;
+                &lt;div class="container-content" id="error"&gt;
+                    &lt;h1&gt;Something went wrong&lt;/h1&gt;
+                    &lt;p&gt;
+                         Authentication failed: &lt;code id="error-code"&gt;&lt;/code&gt;.
+                    &lt;/p&gt;
+                &lt;/div&gt;
+            &lt;/div&gt;
+        &lt;/div&gt;
+        
+        &lt;script language="javascript"&gt;
+            const params = new URLSearchParams(location.search);
+            if (params.has('error')) {
+                document.getElementById('success').style.display = 'none';
+                document.getElementById('error-code').textContent = params.get('error');
+            }
+            else {
+                document.getElementById('error').style.display = 'none';
+            }
+        &lt;/script&gt;
+    &lt;/body&gt;
+&lt;/html&gt;</value>
+  </data>
+  <data name="GoogleSignIn_24" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\GoogleSignIn_24.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="LogoWatermark_180" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\LogoWatermark_180.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
 </root>
