+++ sources/Google.Solutions.IapDesktop.Application/Views/Authorization/AuthorizeView.resx
