+++ sources/Google.Solutions.IapDesktop/Google.Solutions.IapDesktop.csproj
@@ -119,13 +119,6 @@
     <Compile Include="ProcessMitigations.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Windows\AuthorizeView.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="Windows\AuthorizeView.Designer.cs">
-      <DependentUpon>AuthorizeView.cs</DependentUpon>
-    </Compile>
-    <Compile Include="Windows\AuthorizeViewModel.cs" />
     <Compile Include="Windows\ErrorDialog.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -140,11 +133,6 @@
     </Compile>
     <Compile Include="Windows\MainFormViewModel.cs" />
     <Compile Include="Interop\TaskBarList.cs" />
-    <Compile Include="Windows\Resources.Designer.cs">
-      <AutoGen>True</AutoGen>
-      <DesignTime>True</DesignTime>
-      <DependentUpon>Resources.resx</DependentUpon>
-    </Compile>
     <Compile Include="Windows\ToolStripExtensions.cs" />
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
@@ -156,19 +144,11 @@
       <DependentUpon>Resources.resx</DependentUpon>
       <DesignTime>True</DesignTime>
     </Compile>
-    <EmbeddedResource Include="Windows\AuthorizeView.resx">
-      <DependentUpon>AuthorizeView.cs</DependentUpon>
-    </EmbeddedResource>
     <EmbeddedResource Include="Windows\ErrorDialog.resx">
       <DependentUpon>ErrorDialog.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Windows\MainForm.resx">
       <DependentUpon>MainForm.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="Windows\Resources.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <SubType>Designer</SubType>
-      <LastGenOutput>Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
     <None Include="app.manifest" />
     <None Include="packages.config" />
