+++ sources/Google.Solutions.IapDesktop/Properties/Resources.resx
@@ -121,14 +121,8 @@
   <data name="AddUser_16" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\AddUser_16.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
-  <data name="GoogleSignIn_24" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\GoogleSignIn_24.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
   <data name="ConsentScreen" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\ConsentScreen.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="LogoWatermark_180" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\LogoWatermark_180.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
   <data name="Documentation_16" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\Documentation_16.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
@@ -138,9 +132,6 @@
   </data>
   <data name="UserWhite_16" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\UserWhite_16.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="AccentGradient_450" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\AccentGradient_450.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
   <data name="EndpointVerificationWhite_16" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\EndpointVerificationWhite_16.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
